requirements.txt
aiohttp
astroid
async-timeout
attrs
certifi
cffi
chardet
chicken
colorama
discord.py
emoji
feedparser
idna
imgurpython
isort
lazy-object-proxy
mccabe
multidict
pbr
pycparser
pylint
PyNaCl
requests
six
tendo
typed-ast
urbandictionary
urllib3
websockets
wrapt
yarl